training/requirements.txt
imgaug==0.4.0
natsort==6.0.0
numpy==1.19.5
Pillow==6.1.0
pytorch-lightning==1.2.4
scikit-image==0.17.2
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.11.0
sklearn==0.0
statsmodels==0.12.1
torch==1.7.0
torchmetrics==0.3.2
torchvision==0.4.0
tqdm==4.59.0
wfdb==3.1.1
